replace old DateField with new DateField
