+++ packages/bento-design-system/src/DateField/Config.ts
@@ -4,6 +4,7 @@ import { BodyProps } from "../Typography/Body/Body";
 import { LabelProps } from "../Typography/Label/Label";
 import { BorderRadiusConfig } from "../util/BorderRadiusConfig";
 import { Children } from "../util/Children";
+import { vars } from "../vars.css";
 
 export type DateFieldConfig = {
   radius: BorderRadiusConfig;
@@ -17,8 +18,8 @@ export type DateFieldConfig = {
     open: (props: IconProps) => Children;
     close: (props: IconProps) => Children;
   };
-  dayWidth: number;
-  dayHeight: number;
+  dayWidth: keyof typeof vars.space;
+  dayHeight: keyof typeof vars.space;
   dayRadius: BorderRadiusConfig;
   daySize: BodyProps["size"];
 };
