+++ packages/bento-design-system/src/DateField/Input.tsx
@@ -1,144 +1,172 @@
-import { useTextField } from "@react-aria/textfield";
-import { InputHTMLAttributes, RefObject, useEffect, useState } from "react";
-import { Box } from "..";
-import { input } from "./DateField.css";
-import { getInputValue, parseDate as _parseDate } from "@datepicker-react/hooks";
-import { useDateFormatter } from "@react-aria/i18n";
-import InputMask from "react-input-mask";
-import { DateFormatter } from "@internationalized/date";
+import { CalendarDate, createCalendar } from "@internationalized/date";
+import { AriaDateFieldOptions, useDateField, useDateSegment } from "@react-aria/datepicker";
+import { useLocale } from "@react-aria/i18n";
+import {
+  DateFieldState,
+  DateSegment as DateSegmentType,
+  useDateFieldState,
+} from "@react-stately/datepicker";
+import { RefObject, useRef } from "react";
+import { Box } from "../Box/Box";
+import { inputRecipe } from "../Field/Field.css";
+import { bodyRecipe } from "../Typography/Body/Body.css";
+import { useBentoConfig } from "../BentoConfigContext";
+import { getRadiusPropsFromConfig } from "../util/BorderRadiusConfig";
+import { Body } from "../Typography/Body/Body";
+import { Inline } from "../Layout/Inline";
+import useDimensions from "react-cool-dimensions";
+import { IconCalendar, IconMinus } from "../Icons";
+import { AriaButtonProps } from "@react-types/button";
+import { IconButton } from "../IconButton/IconButton";
+import { getReadOnlyBackgroundStyle } from "../Field/utils";
+import { match, __ } from "ts-pattern";
+import { dateFieldRecipe, dateSegment } from "./DateField.css";
+import { ValidationState } from "@react-types/shared";
 
-type Props = {
-  for: "startDate" | "endDate";
-  currentDate: Date | null;
-  inputRef: RefObject<HTMLInputElement>;
-  isFocused: boolean;
-  onClick: () => void;
-  onFocus: () => void;
-  onBlur: () => void;
-  onChange: (date: Date) => void;
-  disabled: boolean;
-  readOnly: boolean;
-  onDateSelect: (date: Date) => void;
-  onDateClear: () => void;
-  isOpen: boolean;
+type Props = (
+  | { type: "single"; fieldProps: AriaDateFieldOptions<CalendarDate> }
+  | {
+      type: "range";
+      fieldProps: {
+        start: AriaDateFieldOptions<CalendarDate>;
+        end: AriaDateFieldOptions<CalendarDate>;
+      };
+    }
+) & {
+  buttonProps: AriaButtonProps<"button">;
+  ref: RefObject<HTMLInputElement>;
 };
 
-function getDatePattern(dateFormatter: DateFormatter): string {
-  const parts = dateFormatter.formatToParts(new Date());
+function DateSegment({ segment, state }: { segment: DateSegmentType; state: DateFieldState }) {
+  const ref = useRef(null);
+  const { segmentProps } = useDateSegment(segment, state, ref);
+  const config = useBentoConfig().input;
 
-  return parts
-    .map((part) => {
-      switch (part.type) {
-        case "day":
-          return "dd";
-        case "month":
-          return "MM";
-        case "year":
-          return "yyyy";
-        default:
-          return undefined;
-      }
-    })
-    .filter((a) => !!a)
-    .join("/");
+  return (
+    <Box {...segmentProps} ref={ref} className={dateSegment} readOnly={state.isReadOnly}>
+      <Body size={config.fontSize} color="inherit">
+        {segment.text}
+      </Body>
+    </Box>
+  );
 }
 
-function parseDate(value: string, pattern: string): Date | null {
-  const parsedDate = _parseDate(value, pattern, new Date());
-  if (!isNaN(parsedDate.getTime())) {
-    return parsedDate;
-  }
-  return null;
+function DateField({ fieldProps }: { fieldProps: AriaDateFieldOptions<CalendarDate> }) {
+  const { locale } = useLocale();
+  const ref = useRef(null);
+  const state = useDateFieldState({
+    ...fieldProps,
+    locale,
+    createCalendar,
+  });
+  const { fieldProps: dateFieldProps } = useDateField(fieldProps, state, ref);
+  return (
+    <Box {...dateFieldProps} ref={ref}>
+      <Inline space={0}>
+        {state.segments.map((segment, i) => (
+          <DateSegment key={i} segment={segment} state={state} />
+        ))}
+      </Inline>
+    </Box>
+  );
 }
 
 export function Input(props: Props) {
-  const dateFormatter = useDateFormatter({ day: "2-digit", month: "2-digit", year: "numeric" });
-  const datePattern = getDatePattern(dateFormatter);
-  const dateMask = datePattern.replace(/[a-zA-Z]/g, "9");
-  const [value, setValue] = useState("");
-  function setCurrentValue() {
-    return props.currentDate
-      ? setValue(getInputValue(props.currentDate, (date) => dateFormatter.format(date), ""))
-      : setValue("");
-  }
-  const onFocus = () => {
-    if (!props.readOnly) {
-      props.inputRef.current && props.inputRef.current.select();
-      props.onFocus();
-    }
-  };
-  const onBlur = () => {
-    props.onBlur();
-    setCurrentValue();
-  };
+  const config = useBentoConfig().input;
 
-  const { inputProps } = useTextField(
-    {
-      ...props,
-      label: props.for,
-      placeholder: datePattern.toUpperCase(),
-      type: "text",
-      value,
-      onChange: (value) => {
-        props.onClick();
-        setValue(value);
-        const newDate = parseDate(value, datePattern);
-        if (newDate) {
-          props.onChange(newDate);
-        }
-      },
-      isDisabled: props.disabled,
-      isReadOnly: props.readOnly,
-      onBlur,
-      onFocus,
-      onKeyDown: (e) => {
-        if (e.key === "Enter") {
-          if (props.isOpen) {
-            const date = parseDate(value, datePattern);
-            if (date) {
-              props.onDateSelect(date);
-            } else {
-              if (value === "") {
-                props.onDateClear();
-              } else {
-                setCurrentValue();
-              }
-            }
-          } else {
-            props.onClick();
-          }
-        }
-      },
-      autoComplete: "off",
-    },
-    props.inputRef
-  );
+  const { observe: rightAccessoryRef, width: rightAccessoryWidth } = useDimensions({
+    // This is needed to include the padding in the width
+    useBorderBoxSize: true,
+  });
 
-  useEffect(setCurrentValue, [props.currentDate]);
+  const { validationState, isDisabled, isReadOnly } = match(props)
+    .with({ type: "single" }, (props) => {
+      return {
+        validationState: props.fieldProps.isReadOnly
+          ? "notSet"
+          : props.fieldProps.validationState ?? "notSet",
+        isDisabled: props.fieldProps.isDisabled,
+        isReadOnly: props.fieldProps.isReadOnly,
+      } as const;
+    })
+    .with({ type: "range" }, (props) => {
+      const isReadOnly = props.fieldProps.start.isReadOnly && props.fieldProps.end.isReadOnly;
+      const validationState: ValidationState | "notSet" = isReadOnly
+        ? ("notSet" as const)
+        : match([
+            props.fieldProps.start.validationState,
+            props.fieldProps.end.validationState,
+          ] as const)
+            .with(["invalid", __], [__, "invalid"], () => "invalid" as const)
+            .with([undefined, undefined], () => "notSet" as const)
+            .with([__, "valid"], ["valid", __], () => "valid" as const)
+            .exhaustive();
+
+      return {
+        validationState,
+        isDisabled: props.fieldProps.start.isDisabled && props.fieldProps.end.isDisabled,
+        isReadOnly,
+      } as const;
+    })
+    .exhaustive();
 
   return (
-    <InputMask {...inputProps} mask={dateMask} maskChar="">
-      {(maskedInputProps: InputHTMLAttributes<HTMLInputElement>) => {
-        const { onFocus, onBlur, ...rest } = inputProps;
-        return (
-          <Box
-            as="input"
-            type="text"
-            className={input}
-            ref={props.inputRef}
-            onClick={() => {
-              if (!props.disabled && !props.readOnly) {
-                props.onClick();
-              }
-            }}
-            {...rest}
-            {...maskedInputProps}
-            color={undefined}
-            width={undefined}
-            height={undefined}
+    <Box
+      {...getRadiusPropsFromConfig(config.radius)}
+      paddingX={config.paddingX}
+      paddingY={config.paddingY}
+      background={config.background.default}
+      className={[
+        inputRecipe({ validation: validationState }),
+        bodyRecipe({
+          color: isDisabled ? "disabled" : "primary",
+          weight: "default",
+          size: config.fontSize,
+          ellipsis: false,
+        }),
+        dateFieldRecipe({ validation: validationState }),
+        {
+          readOnly: isReadOnly,
+        },
+      ]}
+      style={{ paddingRight: rightAccessoryWidth, ...getReadOnlyBackgroundStyle(config) }}
+      position="relative"
+      disabled={isDisabled}
+      readOnly={isReadOnly}
+    >
+      {props.type === "single" ? (
+        <DateField fieldProps={props.fieldProps} />
+      ) : (
+        <Inline space={16}>
+          <DateField fieldProps={props.fieldProps.start} />
+          <IconMinus size={24} />
+          <DateField fieldProps={props.fieldProps.end} />
+        </Inline>
+      )}
+      {!isReadOnly && (
+        <Box
+          ref={rightAccessoryRef}
+          position="absolute"
+          display="flex"
+          justifyContent="center"
+          alignItems="center"
+          paddingX={config.paddingX}
+          top={0}
+          bottom={0}
+          right={0}
+        >
+          <IconButton
+            kind="transparent"
+            hierarchy="secondary"
+            label="Calendar"
+            size={16}
+            icon={IconCalendar}
+            {...props.buttonProps}
+            onPress={props.buttonProps.onPress!}
+            isDisabled={isDisabled}
           />
-        );
-      }}
-    </InputMask>
+        </Box>
+      )}
+    </Box>
   );
 }
