+++ packages/bento-design-system/src/index.ts
@@ -26,7 +26,7 @@ export * from "./Chip/Chip";
 export * from "./ContentBlock/ContentBlock";
 export * from "./ContentWithSidebar/ContentWithSidebar";
 export * from "./DateField/DateField";
-export * from "./DateField2/Calendar";
+export * from "./DateField/Calendar";
 export * from "./Disclosure/Disclosure";
 export * from "./DisclosureGroup/DisclosureGroup";
 export * from "./Divider/Divider";
